Revert "sstatesig: include native/cross/nativesdk deps in target signatures"
This reverts commit 9abc0028fc2b8d7d79b280d30bdbd987fca2b6d8.

In FSL SDK, cache iso was delivered to speed up build. It is needless
and safe not to include native/cross/nativesdk deps in target
signatures.
using extend sstate_rundepfilter is a way mentioned in the
commit, but a control variant to enable/disable it may be another
better option. will submit patch later. revert it for sdk release now.

Change-Id: Ie589df71986b4335b98c6f75c251872954db73cd
Reviewed-on: http://git.am.freescale.net:8181/12759
Reviewed-by: Ting Liu <ting.liu@freescale.com>
Tested-by: Ting Liu <ting.liu@freescale.com>